Refactor and update meson.build files, fix includes, and add error tests
- Refactor include directories
- Added new test file `test/test_cmp_errors.c`
- Improved test names
