+++ examples/meson.build
@@ -1,15 +1,14 @@
-exe_multi= executable('multi_pass_compression',
+exe_multi = executable('multi_pass_compression',
   'multi_pass_compression.c',
   link_with : cmp_lib,
-  include_directories: inc)
+  include_directories : cmp_inc)
 
-test('Run multi-pass compression example',
-  exe_multi)
+test('Run multi-pass compression example', exe_multi)
 
-exe_single= executable('single_pass_compression',
+
+exe_single = executable('single_pass_compression',
   'single_pass_compression.c',
   link_with : cmp_lib,
-  include_directories: inc)
+  include_directories : cmp_inc)
 
-test('Run single-pass compression example',
-  exe_single)
+test('Run single-pass compression example', exe_single)
