+++ lib/meson.build
@@ -1,9 +1,11 @@
 subdir('common')
 subdir('compress')
 
-inc = include_directories('.')
+cmp_inc = include_directories('.')
+common_inc = include_directories('common')
+compress_inc = include_directories('compress')
 
 cmp_lib = static_library('cmp',
   src_common, src_compress,
-  include_directories: inc,
+  include_directories: cmp_inc,
   install: true)
